ds18b20 and one wire drivers restyled

+++ STM32F0xx/adc.c
@@ -8,9 +8,9 @@ uint32_t ADCx_GetCalibrationFactor (ADC_TypeDef * ADCx)
 	return ADCx->DR;
 }
 
-float GetIntTemp (void)
+double GetIntTemp (void)
 {
-	float res;
+	double res;
 	
 	ADC1->CHSELR = ADC_CHSELR_CHSEL16;
 	ADC1->SMPR = ADC_SMPR_SMP;
@@ -18,7 +18,8 @@ float GetIntTemp (void)
 	ADC1->CR = ADC_CR_ADSTART | ADC_CR_ADEN;
 	while (!(ADC1->ISR & ADC_ISR_EOC));
 	
-	res = (110.0 - 30.0) * (float)((int16_t)ADC1->DR - TS_CAL1) / (float)(TS_CAL2 - TS_CAL1) + 30.0;
+	res = (110.0 - 30.0) * (double)((int16_t)ADC1->DR - TS_CAL1)
+		/ (double)(TS_CAL2 - TS_CAL1) + 30.0;
 	
 	ADC->CCR &= ~ADC_CCR_TSEN;
 	ADC1->CR = 0;
@@ -26,9 +27,9 @@ float GetIntTemp (void)
 	return res;
 }
 
-float GetIntVDD (void)
+double GetIntVDD (void)
 {
-	float res;
+	double res;
 	
 	ADC1->CHSELR = ADC_CHSELR_CHSEL17;
 	ADC1->SMPR = ADC_SMPR_SMP;
@@ -36,7 +37,7 @@ float GetIntVDD (void)
 	ADC1->CR = ADC_CR_ADSTART | ADC_CR_ADEN;
 	while (!(ADC1->ISR & ADC_ISR_EOC));
 	
-	res = 3.3 * (float)VREFINT_CAL / (float)ADC1->DR;
+	res = 3.3 * (double)VREFINT_CAL / (double)ADC1->DR;
 	
 	ADC->CCR &= ~ADC_CCR_VREFEN;
 	ADC1->CR = 0;
